rename: 보다 직관적인 MockChatData로 파일명 수정
